modelling of valid
