+++ server/models/contactUsModel.js
@@ -1,7 +1,5 @@
 const mongoose = require('mongoose');
 
-const joi = require('joi');
-
 const Schema = mongoose.Schema();
 
 const contactUsSchema = new Schema({
@@ -11,12 +9,3 @@ const contactUsSchema = new Schema({
 });
 
 module.exports = mongoose.model('contactUs', contactUsSchema);
-
-exports.validContactUs = (bodyData) => {
-  const joiSchema = joi.object({
-    email: joi.String().joi.min().joi.max().joi.require(),
-    inquiry: joi.String().joi.min().joi.max().joi.require(),
-    content: joi.String().joi.min().joi.max().joi.require(),
-  });
-  return joiSchema.validate(bodyData);
-};
